Disable research page and remove navbar link
